Toger#Rebuild statistic part, and invoke pbt in marble
