+++ model/type.go
@@ -14,3 +14,13 @@ type OnlineUser struct {
 	User      []*Person
 	LoginTime string
 }
+
+// Statistics statitic struct for handler statistics
+type Statistic struct {
+	Numbers []float64
+	Count   int
+	Mean    float64
+	Median  float64
+	ErrMsg  string
+	Get     bool
+}
