+++ template/list.htm
@@ -24,16 +24,50 @@
   <h2><a href="./list">List Example Home</a></h2>
 <table>
     <tr>
-        <td class="docs">
-        <p>Golang template</p>
-        </td>
-        <td class="code empty leading">
-        <a href="/tpl">Generate from template</a>
+        <td class="docs" colSpan="2">
+        <p>General golang HTTP</p>
         </td>
     </tr>
     <tr>
         <td class="docs">
-        JSON API example 
+        <p>HTTP header</p> 
+        </td>
+        <td class="code leading">
+        <a href="/test/protocol"> HTTP Protocol</a>
+        </td>
+    </tr>
+    <tr>
+        <td class="docs">
+        <p>From template</p>
+        </td>
+        <td class="code empty leading">
+        <a href="/tpl">Template</a>
+        </td>
+    </tr>
+    <tr>
+        <td class="docs">
+        <p>Statistic(none template)</p>
+        </td>
+        <td class="code leading">
+        <a href="/statistic">Statistic</a>
+        </td>
+    </tr>
+    <tr>
+        <td class="docs">
+        <p>Advanced Statistic(with template)</p> 
+        </td>
+        <td class="code leading">
+        <a href="/statistics">Advanced Statistic</a>
+        </td>
+    </tr>
+    <tr>
+        <td class="docs" colSpan="2">
+        <p>API with different protocol</p>
+        </td>
+    </tr>
+    <tr>
+        <td class="docs">
+        <p>JSON example</p>
         </td>
         <td class="code leading">
         <a href="/json">JSON API</a>
@@ -41,10 +75,10 @@
     </tr>
     <tr>
         <td class="docs">
-        HTTP header test 
+        <p>Protobuf test</p>
         </td>
         <td class="code leading">
-        <a href="/test/protocol"> HTTP Protocol Test</a>
+        <a href="/pbt">Protobuf test</a>
         </td>
     </tr>
 </table>
